requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
psycopg2==2.6
requests==2.7.0
SQLAlchemy==1.0.4
Werkzeug==0.10.4
